feat: refactored encrypted data on sign-in
