+++ utils/API/index.js
@@ -1,5 +1,6 @@
 export * from "./amount";
 export * from "./decryptUserData";
+export * from "./encryptUserData";
 export * from "./response";
 export * from "./skip";
 export * from "./take";
